Card #253: add latest observation date and value to alert table SQL, save alert status when importing, todo: email alerts, hide notification Migration: flag for delete existing patients during import (very expensive)

+++ root/api/src/main/java/org/patientview/api/service/impl/MigrationServiceImpl.java
@@ -93,6 +93,7 @@ public class MigrationServiceImpl extends AbstractServiceImpl<MigrationServiceIm
     private DataSource dataSource;
 
     private static final String COMMENT_RESULT_HEADING = "resultcomment";
+    private static final boolean DELETE_EXISTING = true;
 
     public Long migrateUser(MigrationUser migrationUser)
             throws EntityExistsException, ResourceNotFoundException, MigrationException {
@@ -106,13 +107,18 @@ public class MigrationServiceImpl extends AbstractServiceImpl<MigrationServiceIm
 
         org.patientview.api.model.User apiUser = userService.getByUsername(user.getUsername());
 
-        // delete user if already exists
+        // delete user if already exists (expensive, not to be used for live migration)
         if (apiUser != null) {
-            try {
-                userService.delete(apiUser.getId());
-            } catch (ResourceForbiddenException | FhirResourceException e) {
-                LOG.error("Cannot delete user with id " + apiUser.getId());
-                throw new MigrationException(e);
+            if (DELETE_EXISTING) {
+                try {
+                    userService.delete(apiUser.getId());
+                } catch (ResourceForbiddenException | FhirResourceException e) {
+                    LOG.error("Cannot delete user with id " + apiUser.getId());
+                    throw new MigrationException(e);
+                }
+            } else {
+                userId = apiUser.getId();
+                userMigration = userMigrationService.getByPatientview2Id(userId);
             }
         }
